requirements.txt
asgiref==3.11.1
Django==6.0.3
django-appconf==1.2.0
django-axes==8.3.1
django-environ==0.13.0
django-imagekit==6.1.0
django-ratelimit==4.1.0
pilkit==3.0
pillow==12.1.1
sqlparse==0.5.5
